+++ projects/igniteui-angular/src/lib/date-picker/date-picker.component.spec.ts
@@ -44,10 +44,10 @@ describe('IgxDatePicker', () => {
         });
 
         it('Initialize a datepicker component with id', () => {
-            const domDatePicker = fixture.debugElement.query(By.css('igx-datePicker')).nativeElement;
+            const domDatePicker = fixture.debugElement.query(By.css('igx-date-picker')).nativeElement;
 
-            expect(datePicker.id).toContain('igx-datePicker-');
-            expect(domDatePicker.id).toContain('igx-datePicker-');
+            expect(datePicker.id).toContain('igx-date-picker-');
+            expect(domDatePicker.id).toContain('igx-date-picker-');
 
             datePicker.id = 'customDatePicker';
             fixture.detectChanges();
@@ -126,7 +126,7 @@ describe('IgxDatePicker', () => {
         });
 
         it('Handling keyboard navigation with `space`(open) and `esc`(close) buttons', fakeAsync(() => {
-            const datePickerDom = fixture.debugElement.query(By.css('igx-datepicker'));
+            const datePickerDom = fixture.debugElement.query(By.css('igx-date-picker'));
             let overlayToggle = document.getElementsByTagName('igx-toggle');
             expect(overlayToggle.length).toEqual(0);
 
@@ -147,7 +147,7 @@ describe('IgxDatePicker', () => {
 
         it('When datepicker is closed and the dialog disappear the focus should remain on the input',
         fakeAsync(() => {
-            const datePickerDom = fixture.debugElement.query(By.css('igx-datepicker'));
+            const datePickerDom = fixture.debugElement.query(By.css('igx-date-picker'));
             let overlayToggle = document.getElementsByTagName('igx-toggle');
             expect(overlayToggle.length).toEqual(0);
 
@@ -274,7 +274,7 @@ describe('IgxDatePicker', () => {
 
 @Component({
     template: `
-        <igx-datePicker [formatter]="customFormatter" [value]=date></igx-datePicker>
+        <igx-date-picker [formatter]="customFormatter" [value]=date></igx-date-picker>
     `
 })
 export class IgxDatePickerWithCustomFormatterComponent {
@@ -288,7 +288,7 @@ export class IgxDatePickerWithCustomFormatterComponent {
 
 @Component({
     template: `
-        <igx-datePicker [value]="date" [weekStart]="1"></igx-datePicker>
+        <igx-date-picker [value]="date" [weekStart]="1"></igx-date-picker>
     `
 })
 export class IgxDatePickerWithWeekStartComponent {
@@ -298,7 +298,7 @@ export class IgxDatePickerWithWeekStartComponent {
 
 @Component({
     template: `
-        <igx-datePicker [labelVisibility]="labelVisibility"></igx-datePicker>
+        <igx-date-picker [labelVisibility]="labelVisibility"></igx-date-picker>
     `
 })
 export class IgxDatePickerTestComponent {
@@ -309,7 +309,7 @@ export class IgxDatePickerTestComponent {
 
 @Component({
     template: `
-        <igx-datePicker [value]="date" [formatOptions]="formatOptions"></igx-datePicker>
+        <igx-date-picker [value]="date" [formatOptions]="formatOptions"></igx-date-picker>
     `
 })
 export class IgxDatePickerWithPassedDateComponent {
@@ -325,7 +325,7 @@ export class IgxDatePickerWithPassedDateComponent {
 
 @Component({
     template: `
-        <igx-datePicker [value]="date" [locale]="'de-DE'"></igx-datePicker>
+        <igx-date-picker [value]="date" [locale]="'de-DE'"></igx-date-picker>
     `
 })
 export class IgxDatePickerWIthLocaleComponent {
@@ -335,7 +335,7 @@ export class IgxDatePickerWIthLocaleComponent {
 
 @Component({
     template: `
-        <igx-datePicker [(ngModel)]="val"></igx-datePicker>
+        <igx-date-picker [(ngModel)]="val"></igx-date-picker>
     `
 })
 export class IgxDatePickerNgModelComponent {